restructure redux action types
