+++ src/redux/actions/index.js
@@ -1,3 +1,11 @@
+// Types (the Heba method)
+export { SET_AUTHORS, ADD_AUTHOR, SET_BOOKS, ADD_BOOK } from "./actionTypes";
+
+// Authors
 export { fetchAuthors, postAuthor } from "./authors";
+
+// Books
 export { fetchBooks, postBook } from "./books";
+
+// Authentication
 export { login, logout, signup } from "./authentication";
